Struct block_game_clone_backend::board::Board

pub struct Board {
    pub cells: [[[BoardCell; 8]; 8]; 8],
    pub height_limits: Vec<Vec<usize>>,
    pub center: Vector3<f32>,
}

Represents the state of the board

Fields
cells: [[[BoardCell; 8]; 8]; 8]height_limits: Vec<Vec<usize>>

used to construct, show available space to player

center: Vector3<f32>

useful for centering a camera

Implementations

impl Board

fn map_to_heights(map: TwoPlayerMap) -> Vec<Vec<usize>>

pub fn new(game_mode: GameMode) -> Self

pub fn calculate_score(&self) -> HashMap<Player, i8>

pub fn get_available_positions(&self) -> Vec<Vector3<f32>>

returns all available positions on the board. Used for searching for possible moves

fn get_highest_player(&self, x: i8, z: i8, y_max: i8) -> Option<Player>

Returns the highest player for a given column, used for scoring

fn get_from_index(&self, (x, y, z): (usize, usize, usize)) -> Option<&BoardCell>

fn get_from_index_mut( &mut self, (x, y, z): (usize, usize, usize) ) -> Option<&mut BoardCell>

pub fn get(&self, index: Vector3<f32>) -> Option<&BoardCell>

pub fn get_mut(&mut self, index: Vector3<f32>) -> Option<&mut BoardCell>

pub fn add_cubes(&mut self, cubes: &Vec<Cube>)

pub fn supports(&self, position: &Vector3<f32>) -> bool

fn get_adjacent_cells(&self, position: &Vector3<f32>) -> Vec<&BoardCell>

pub fn check_touches_piece( &self, new_piece_coords: Vec<Cube> ) -> Result<Vec<Cube>, Vec<Cube>>

implements game rule requiring players to play touching their own piece.

The first player may play anywhere. The second must touch the first

fn player_has_played(&self, player: Player) -> Option<Player>

determines whether or not a player has played

pub fn check_in_bounds_no_collision( &self, position: Vector3<f32> ) -> Option<CubeError>

fn new_board_from_2d_heights(heights_2d: Vec<Vec<usize>>) -> Self