Struct block_game_clone_backend::piece::Piece

pub struct Piece {
    pub coords: Vec<Vector3<f32>>,
}

Represents a piece as a Vec of offsets as nalgebra::Vector3<f32> from [0,0,0]

Fields
coords: Vec<Vector3<f32>>
Implementations

impl Piece

pub fn get_moved_copy(&self, position: Vector3<f32>) -> Piece

applies vector translation to constituent cubes

pub fn set_origin(&mut self, new_origin: Vector3<f32>)

Imagine picking up a Polycube by one of the cubes

The position of the cubes are expressed as offsets from this selected cube which is at the origin

pub fn rotate(&mut self, rotation_axis: RotationAxis)

Rotates the piece about the input RotationAxis

fn apply_rotation(&self, rotation: Rotation3<f32>) -> Self

pub fn from_vec_i8_array(i: Vec<[i8; 3]>) -> Piece

pub fn supports(&self, position: &Vector3<f32>) -> bool

Does this piece support a given position. All cubes must be supported. A cube can be supported by another cube in the same piece

pub fn get_available_piece_rotations(&self) -> Vec<Piece>

Give all possible rotations of the current piece